feat(nocgo): enable to build with CGO_ENABLED=0
