+++ exploit/exploit.go
@@ -1,3 +1,5 @@
+// +build !scanner
+
 package exploit
 
 import (
