+++ gost/pseudo.go
@@ -1,8 +1,8 @@
+// +build !scanner
+
 package gost
 
 import (
-	"strings"
-
 	"github.com/future-architect/vuls/models"
 	"github.com/knqyf263/gost/db"
 )
@@ -15,8 +15,4 @@ type Pseudo struct {
 // DetectUnfixed fills cve information that has in Gost
 func (pse Pseudo) DetectUnfixed(driver db.DB, r *models.ScanResult, _ bool) (int, error) {
 	return 0, nil
-}
-
-func major(osVer string) (majorVersion string) {
-	return strings.Split(osVer, ".")[0]
 }
