+++ msf/msf.go
@@ -1,3 +1,5 @@
+// +build !scanner
+
 package msf
 
 import (
