+++ oval/oval.go
@@ -1,3 +1,5 @@
+// +build !scanner
+
 package oval
 
 import (
