+++ report/cve_client.go
@@ -1,3 +1,5 @@
+// +build !scanner
+
 package report
 
 import (
